Move courts.py to new fixtures Python module
